lab2 Finished debugging exercises and IntList implementation

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,8 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        //fix: used to be int here and no cast to float for division
+        float quotient = ((float) top) / bottom;
         int result = Math.round(quotient);
         return result;
     }
